Mechanism for finding project root, info command.
This adds a mechanism for finding the project root, which is denoted
by the ".glue_and_stickers" file.  This file is a simple marker, and
does not need to contain any information, though it might be a handy
place to store metadata that wouldn't otherwise just go in a config
file.

Next up: asset management!