Added fancy PageAccessor class.
The PageAccessor class provides basic access to the metadata
associated to discrete pages.

The PageAccessor class holds a weakref to PageData, so as to require
multiple PageAccessor classes to work on the same PageData object if
they have the same page root path.

This provides functionality like whether or not something is
"published", and a nice pythonic API for reading and manipulating the
page metadata.