More reasonable Page implementation, queries, info command.
The "Page" class shouldn't be used directly, but rather the
"access_page" method should be used instead.

This will ensure that only one Page object is ever created for a given
page path, and simplifies things quite a bit.

This commit also adds some utility functions for querying pages, which
are found in assets.py.

The info command has been enhanced to show an overview of all pages
when invoked from a non-page directory in the project, and shows a
more detailed summary of a page when invoked from a page directory.

This also adds a check for the python version, and errors if python2
is used, as it is unsupported.